Use component for json-editor and other review fixes
