+++ app/assets/javascripts/discourse/app/controllers/json-schema-editor.js
@@ -1,56 +1,13 @@
+import { action } from "@ember/object";
 import Controller from "@ember/controller";
-import { create } from "virtual-dom";
-import { iconNode } from "discourse-common/lib/icon-library";
-import loadScript from "discourse/lib/load-script";
 import ModalFunctionality from "discourse/mixins/modal-functionality";
-import { schedule } from "@ember/runloop";
 
 export default Controller.extend(ModalFunctionality, {
   editor: null,
-  jsonSchema: null,
 
-  onShow() {
-    loadScript("/javascripts/jsoneditor.js").then(() => {
-      schedule("afterRender", () => {
-        const el = document.querySelector("#json-editor-holder");
-
-        /* eslint-disable no-undef */
-        JSONEditor.defaults.options.theme = "bootstrap4";
-
-        JSONEditor.defaults.iconlibs = {
-          discourseIcons: DiscourseJsonSchemaEditorIconlib,
-        };
-        JSONEditor.defaults.options.iconlib = "discourseIcons";
-
-        this.editor = new JSONEditor(el, {
-          schema: this.model.jsonSchema,
-          disable_array_delete_all_rows: true,
-          disable_array_delete_last_row: true,
-          disable_array_reorder: true,
-          disable_array_copy: true,
-          disable_edit_json: true,
-          disable_properties: true,
-          disable_collapse: true,
-          startval: this.model.value ? JSON.parse(this.model.value) : null,
-        });
-        /* eslint-enable no-undef */
-      });
-    });
+  @action
+  saveChanges(value) {
+    this.set("model.value", value);
+    this.send("closeModal");
   },
 });
-
-class DiscourseJsonSchemaEditorIconlib {
-  constructor() {
-    this.mapping = {
-      delete: "times",
-      add: "plus",
-    };
-  }
-
-  getIcon(key) {
-    if (!this.mapping[key]) {
-      return;
-    }
-    return create(iconNode(this.mapping[key]));
-  }
-}
