+++ app/assets/javascripts/discourse/app/templates/components/json-editor.hbs
@@ -1,0 +1,11 @@
+{{#d-modal-body rawTitle=(i18n "admin.site_settings.json_schema.modal_title" name=model.settingName)}}
+  <div id="json-editor-holder"></div>
+{{/d-modal-body}}
+
+<div class="modal-footer">
+  {{d-button
+    class="btn-primary"
+    action=(action "saveChanges")
+    label="save"
+  }}
+</div>
