+++ app/assets/javascripts/discourse/app/templates/modal/json-schema-editor.hbs
@@ -1,11 +1,1 @@
-{{#d-modal-body rawTitle=(i18n "admin.site_settings.json_schema.modal_title" name=model.settingName)}}
-  <div id="json-editor-holder"></div>
-{{/d-modal-body}}
-
-<div class="modal-footer">
-  {{d-button
-    class="btn-primary"
-    action=(action "closeModal")
-    label="admin.site_settings.json_schema.close_editor"
-  }}
-</div>
+{{json-editor model=this.model saveChangesCallback=(action "saveChanges") tagName=""}}
